+++ chatmaild/src/chatmaild/filtermail.py
@@ -38,6 +38,12 @@ def check_openpgp_payload(payload: bytes):
 
         packet_type_id = payload[i] & 0x3F
         i += 1
+
+        while payload[i] >= 224 and payload[i] < 255:
+            # Partial body length.
+            partial_length = 1 << (payload[i] & 0x1F)
+            i += 1 + partial_length
+
         if payload[i] < 192:
             # One-octet length.
             body_len = payload[i]
@@ -56,7 +62,7 @@ def check_openpgp_payload(payload: bytes):
             )
             i += 5
         else:
-            # Partial body length is not allowed.
+            # Impossible, partial body length was processed above.
             return False
 
         i += body_len